requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.8.2
httpx==0.27.2
supabase==2.6.0
python-multipart==0.0.9
langchain-text-splitters==0.2.4
pdfplumber==0.11.4
python-docx==1.1.2
unidecode==1.3.8
langchain-openai==0.1.3